@@ -3,15 +3,15 @@
   "productName": "trojan-electron",
   "version": "1.0.0",
   "description": "My Electron application description",
-  "main": "src/main",
+  "main": "main",
   "scripts": {
-    "main:dev": "electron .",
-    "main:build": "vite build --config ./vite.main.config.ts",
-    "view:dev": "vite",
-    "view:preview": "vite preview",
-    "view:build": "tsc && vite build",
-    "copyview": "cp -rf ./src/view/dist ./.dist/src/view",
-    "packager": "pnpm main:build && pnpm view:build && pnpm copyview && node electron-packager.js",
+    "dev": "electron .",
+    "build": "rollup -c rollup.config.mjs",
+    "dev:view": "vite",
+    "preview:view": "vite preview",
+    "build:view": "tsc && vite build",
+    "cpview": "cp -rf ./view/dist ./.dist/view",
+    "packager": "pnpm build && pnpm build:view && pnpm cpview && node electron-packager.js",
     "prepare": "node ./scripts/prepare.mjs"
   },
   "keywords": [],
@@ -28,7 +28,7 @@
     "@react-hook/throttle": "^2.2.0",
     "@reduxjs/toolkit": "^1.8.1",
     "antd": "^4.20.1",
-    "electron": "^18.2.3",
+    "electron": "23.0.0-alpha.3",
     "electron-log": "^4.4.6",
     "electron-squirrel-startup": "^1.0.0",
     "electron-store": "^8.0.1",
@@ -49,7 +49,9 @@
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "^22.0.0",
+    "@rollup/plugin-json": "^6.0.0",
     "@rollup/plugin-node-resolve": "^13.3.0",
+    "@rollup/plugin-terser": "^0.2.1",
     "@types/ramda": "^0.28.12",
     "@types/react": "^18.0.8",
     "@types/react-dom": "^18.0.0",
@@ -61,6 +63,7 @@
     "jszip": "^3.10.1",
     "mkdirp": "^1.0.4",
     "node-fetch": "^3.2.10",
+    "rollup": "^3.9.1",
     "rollup-plugin-copy": "^3.4.0",
     "standard": "^17.0.0",
     "typescript": "^4.6.3",
