@@ -2,7 +2,7 @@ const fs = require('fs')
 
 const GenerateSchema = require('generate-schema')
 
-const defaultObject = require('../src/common/store-defaults.json')
+const defaultObject = require('../common/store-defaults.json')
 const object = require('./store-object.json')
 
 const schema = GenerateSchema.json('store', [
