@@ -3,7 +3,7 @@ import react from '@vitejs/plugin-react'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  root: './src/view',
+  root: './view',
   base: '',
   plugins: [
     react()
